% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{shiny_beta}
\alias{shiny_beta}
\title{Shiny app for computing and evaluating the alpha and beta parameters of Beta distributions}
\usage{
shiny_beta()
}
\description{
Shiny app for computing and evaluating the alpha and beta parameters of Beta distributions
}
